smoke-backend/requirements-dev.txt
-r requirements.txt
pytest-flask
pytest-factoryboy
factory_boy
pip==10.0.0
bumpversion==0.5.3
wheel==0.30.0
watchdog==0.8.3
flake8==3.5.0
tox==2.9.1
coverage==4.5.1
click==6.7.0
Sphinx==1.7.1
twine==1.10.0
pyyaml==3.12

pytest==3.4.2
pytest-runner==2.11.1
pytest-cov==2.5.1
pluggy==0.6.0
